Refactor round robin tests
Motivation:

The test harness for the round-robin load balancer tests doesn't need to
be so tightly coupled to round-robin tests. It can also be used for the
pick-first load balancer.

Modifications:

- Move the RR LB test into its own file
- Refactor so the test context holds a `LoadBalancer` rather than a
  `RoundRobinLoadBalancer`

Result:

Can be easily extended for pick-first LB tests
